templates can now be saved (experimental); added custom (read: experimental) descriptor; converted robust matcher to read descriptors and keypoints
git-svn-id: https://rarepos.cs.uni-tuebingen.de/svn-path/rabot/trunk/catkin@5227 690284e3-2886-4bbc-99f2-a987503dca67

+++ src/csapex/include/csapex/model/connection.h
@@ -81,7 +81,7 @@ protected:
 
     int id_;
 
-    int message_count;
+    double message_count;
 
     static int next_connection_id_;
 };
